Support Python 2.7.
Note that this also changes the behaviour of to_futhark to perform a
copy if the passed-in array does not have 'C' order.  This is mostly
because Python 2.7's CFFI does not itself raise an error in this case
(Python 3 does), and it was easier (and more practically useful) to do
the copy than to insert our own check.